+++ main.py
@@ -1,13 +1,7 @@
-# -----------------------------
+# -----------------------------------
 # PROJECT NOVA
-# Version 0.0.2
-# -----------------------------
-
-name = "Adeeb"
-college = "CMR University"
-course = "B.Tech CSE Data Science"
-graduation_year = 2027
-dream_job = "AI Engineer"
+# Version 0.0.3
+# -----------------------------------
 
 print("=" * 50)
 print("WELCOME TO PROJECT NOVA")
@@ -15,17 +9,27 @@ print("=" * 50)
 
 print()
 
-print("PROFILE")
-print("-" * 50)
+# User Setup
+name = input("Enter your name: ")
+college = input("Enter your college: ")
+course = input("Enter your course: ")
+graduation = input("Graduation year: ")
+dream_job = input("Dream job: ")
+
+print()
+
+print("=" * 50)
+print("YOUR PROFILE")
+print("=" * 50)
 
 print("Name :", name)
 print("College :", college)
 print("Course :", course)
-print("Graduation :", graduation_year)
+print("Graduation :", graduation)
 print("Dream Job :", dream_job)
 
 print()
 
-print("Project Version : v0.0.2")
+print("Welcome to Project NOVA,", name)
 
 print("=" * 50)