+++ mini-projects/day-02-input-02.py
@@ -1,0 +1,26 @@
+#Build an Interview Bot.
+#The bot should ask:
+#Name
+#Age
+#Favourite Programming Language
+#Dream Company
+#Favourite Sport
+#Favourite Food
+#Then display a summary like an interviewer.
+
+print("="*35)
+print("Candidate Resume")
+name = input("Enter your name :")
+age = input("Enter your age :")
+language = input("Your Favorite Programming Language :")
+company = input("Your Dream Company :")
+sport = input("Your Favorite Sport :")
+food = input("Your favorite Food :")
+print()
+print("="*35)
+print("Candidate Name :",name)
+print("Candidate Age :",age)
+print("Preferred Programming language :",language)
+print("Preferred Company :",company)
+print("Favorite sport and Food :",sport,"and",food)
+print("="*35)