+++ mini-projects/day-04-input.py
@@ -1,0 +1,8 @@
+name = input("Enter your name :")
+city = input("Enter your city :")
+college = input("Enter your college :")
+print("Hello",name)
+print("You live in :",city)
+print("Your college is :",college)
+
+