+++ mini-projects/projects/day03-project-input.py
@@ -1,0 +1,25 @@
+#The program should ask the user:
+#Name
+#Country they want to visit
+#Budget
+#Number of days
+#Favourite activity
+#Then print a "Travel Plan".
+#No copying from me.
+#Design it yourself.
+
+name = input("Enter your name : ")
+country = input("Enter your destination : ")
+budget = input("Enter your budget : ")
+no_of_days = input("Number of days you wanna travel : ")
+fav_activity = input("Your favorite activity : ")
+print()
+print("="*35)
+print("Your Travel Plan")
+print("="*35)
+print("Your Name :",name)
+print("Your Destination Country :",country)
+print("Your Budget :",budget)
+print("How long will you Stay :",no_of_days)
+print("Your Favorite Thing to Do :",fav_activity)
+print("="*35)