+++ Changelog.md
@@ -9,3 +9,10 @@
 - Learned Variables
 - Added User Profile
 - Added Digital ID mini project
+
+## v0.0.3
+
+- Learned input()
+- Added interactive profile setup
+- Added Interview Bot
+- Added Travel Planner